+++ packages/connect/src/hooks/useSocialLink.ts
@@ -3,7 +3,5 @@
 import { useSocialLinkContext } from '../contexts/SocialLink.js'
 
 export const useSocialLink = () => {
-  const { isSocialLinkOpen, waasConfigKey, setIsSocialLinkOpen } = useSocialLinkContext()
-
-  return { isSocialLinkOpen, waasConfigKey, setIsSocialLinkOpen }
+  return useSocialLinkContext()
 }