+++ packages/wallet-widget/src/components/Filter/NetworksFilter.tsx
@@ -10,12 +10,10 @@ export const NetworksFilter = () => {
   const onClickNetwork = (chainId: number) => {
     if (selectedNetworks.length === allNetworks.length) {
       setSelectedNetworks([chainId])
+    } else if (selectedNetworks.includes(chainId)) {
+      setSelectedNetworks(selectedNetworks.filter(n => n !== chainId))
     } else {
-      if (selectedNetworks.includes(chainId)) {
-        setSelectedNetworks(selectedNetworks.filter(n => n !== chainId))
-      } else {
-        setSelectedNetworks([...selectedNetworks, chainId])
-      }
+      setSelectedNetworks([...selectedNetworks, chainId])
     }
   }