+++ packages/wallet-widget/src/components/NetworkBadge.tsx
@@ -1,5 +1,5 @@
-import { getNetwork, getNetworkBackgroundColor, getNetworkColor } from '@0xsequence/connect'
 import { NetworkImage, Text } from '@0xsequence/design-system'
+import { getNetwork, getNetworkBackgroundColor, getNetworkColor } from '@0xsequence/web-sdk-core'
 import React from 'react'
 
 interface NetworkBadgeProps {