+++ packages/wallet-widget/src/hooks/useNavigationHeader.ts
@@ -1,6 +1,5 @@
 import { useNavigationHeaderContext } from '../contexts/NavigationHeader.js'
 
 export const useNavigationHeader = () => {
-  const { search, selectedTab, setSearch, setSelectedTab } = useNavigationHeaderContext()
-  return { search, selectedTab, setSearch, setSelectedTab }
+  return useNavigationHeaderContext()
 }