+++ packages/wallet-widget/src/hooks/useShared.ts
@@ -1,7 +1,5 @@
 import { useSharedContext } from '../contexts/Shared.js'
 
 export const useShared = () => {
-  const { isGuest, signInDisplay } = useSharedContext()
-
-  return { isGuest, signInDisplay }
+  return useSharedContext()
 }